GameInitialization() repairs
My sound function calls in GameInitialization() are adding to the total compute time.  this brought out an issue that was hidden.

GameInitialization() runs parallel to draw().

A variable called LoadTime is used at the TOP of draw()::if(mode=1)  but is not initialized until the BOTTOM of GameInitialization(). Therefore, "mode = 1" needs to be AFTER "LoadTime = 3" to ensure it is initialized before draw() needs it.

Something i could not figure out is why the sound functions need to be called AFTER "mode = 1". It doesnt matter though.  I placed all sound functions at the end of GameInitialization().